+++ packages/gallery/src/components/gallery/proGallery/galleryContainerNew.js
@@ -20,7 +20,6 @@ import { cssScrollHelper } from '../../helpers/cssScrollHelper.js';
 import { createCssLayouts } from '../../helpers/cssLayoutsHelper.js';
 import checkNewGalleryProps from '../../helpers/isNew';
 import VideoScrollHelperWrapper from '../../helpers/videoScrollHelperWrapper.js';
-import { LayoutFixer } from '../../layoutFixer/layoutFixer';
 
 export class GalleryContainer extends React.Component {
   constructor(props) {
@@ -47,7 +46,6 @@ export class GalleryContainer extends React.Component {
       needToHandleShowMoreClick: false,
       gotFirstScrollEvent: false,
       playingVideoIdx: -1,
-      nextVideoIdx: -1,
       viewComponent: null,
     };
 
@@ -825,10 +823,9 @@ export class GalleryContainer extends React.Component {
     }
   }
 
-  setPlayingIdxState(playingVideoIdx, nextVideoIdx) {
+  setPlayingIdxState(playingVideoIdx) {
     this.setState({
       playingVideoIdx,
-      nextVideoIdx,
     });
   }
 
@@ -842,7 +839,7 @@ export class GalleryContainer extends React.Component {
   getScrollCss({ domId, items, container, styleParams }) {
     this.scrollCss = cssScrollHelper.calcScrollCss({
       items,
-      container, 
+      container,
       styleParams,
       domId,
     });
@@ -1046,6 +1043,7 @@ export class GalleryContainer extends React.Component {
         data-key="pro-gallery-inner-container"
         key="pro-gallery-inner-container"
         id={`pro-gallery-inner-container-${this.props.domId}`}
+        className={this.props.isPrerenderMode ? 'pro-gallery-prerender' : ''}
       >
         <ScrollIndicator
           domId={this.props.domId}
@@ -1082,7 +1080,6 @@ export class GalleryContainer extends React.Component {
           customSlideshowInfoRenderer={this.props.customSlideshowInfoRenderer}
           customLoadMoreRenderer={this.props.customLoadMoreRenderer}
           playingVideoIdx={this.state.playingVideoIdx}
-          nextVideoIdx={this.state.nextVideoIdx}
           noFollowForSEO={this.props.noFollowForSEO}
           proGalleryRegionLabel={this.props.proGalleryRegionLabel}
           actions={{
@@ -1125,14 +1122,6 @@ export class GalleryContainer extends React.Component {
             <style dangerouslySetInnerHTML={{ __html: this.dynamicStyles }} />
           )}
         </div>
-        <LayoutFixer
-          layoutFixerBundleUrl={this.props.layoutFixerBundleUrl}
-          layoutFixerScriptType={this.props.layoutFixerScriptType}
-          items={this.items}
-          styles={this.props.styles}
-          domId={this.props.domId}
-          isPrerenderMode={this.props.isPrerenderMode}
-        ></LayoutFixer>
       </div>
     );
   }