+++ packages/lib/src/settings/options/itemShadowOpacityAndColor.js
@@ -3,6 +3,8 @@ import { default as GALLERY_CONSTS } from '../../common/constants';
 
 export default {
   title: 'Item Shadow Color',
+  isRelevantDescription:
+    'First set a Vertical gallery ("Scroll Direction" as "Vertical").\nThen set "Crop Images" to "false" or set "Crop Type" to anything but "Fit".\nThen set "Choose info layout" to "Attached Background" or set "Texts Placement" to "Show On Hover".\nThen set "Enable Item Shadow" to "true".',
   isRelevant: (styleParams) =>
     styleParams.itemEnableShadow &&
     !styleParams.oneRow &&