+++ packages/lib/src/settings/options/textBoxBorderRadius.js
@@ -4,6 +4,8 @@ import { default as GALLERY_CONSTS } from '../../common/constants';
 export default {
   title: 'Text Box Border Radius',
   description: `Set the border radius of the texts container for each item.`,
+  isRelevantDescription:
+    'Set "Texts Placement" to anything but "Show On Hover" and set "Choose info layout" to "Separated Background".',
   isRelevant: (styleParams) =>
     styleParams.titlePlacement !== GALLERY_CONSTS.placements.SHOW_ON_HOVER &&
     styleParams.imageInfoType === GALLERY_CONSTS.infoType.SEPARATED_BACKGROUND,