+++ packages/lib/src/settings/options/textBoxHeight.js
@@ -4,6 +4,8 @@ import { default as GALLERY_CONSTS } from '../../common/constants';
 export default {
   title: 'Text Box Height (pixels)',
   description: `Set the text box Height when the texts are above or below the item`,
+  isRelevantDescription:
+    'First set a Vertical gallery ("Scroll Direction" as "Vertical").\nThen set "Layout Orientation" to "Columns".\nThen set "Max Group Size" to "1".\nThen set "Texts Placement" to "Show Below" or "Show Above".',
   isRelevant: (styleParams) =>
     styleParams.isVertical &&
     styleParams.groupSize === 1 &&