+++ packages/lib/src/settings/options/titlePlacement.js
@@ -4,6 +4,8 @@ import { createOptions } from '../utils/utils';
 
 export default {
   title: 'Texts Placement',
+  isRelevantDescription:
+    'Set a Vertical gallery ("Scroll Direction" as "Vertical"), set "Layout Orientation" to "Columns" and set "Max Group Size" to "1".',
   isRelevant: (styleParams) =>
     styleParams.isVertical &&
     styleParams.groupSize === 1 &&