+++ CHANGELOG.md
@@ -1,3 +1,42 @@
+## [v2.4.2](https://pro-gallery-2-4-2.surge.sh) (21/12/2020)
+ 
+#### OTHER
+ -  improve: added description to isRelevant in settings
+ -  improve: add docs for the isPrerenderMode to the readme
+ -  improve: update enableInfinitScroll settings and changed behavior in the playground
+
+#### GALLERY
+ -  improve (layoutFixer): remove deprecated layoutFixer code
+
+---
+## [v2.4.1](https://pro-gallery-2-4-1.surge.sh) (17/12/2020)
+ 
+#### GALLERY
+ -  improve (prerender): render highres image on prerender and set a prerender class
+
+---
+## [v2.4.0](https://pro-gallery-2-4-0.surge.sh) (16/12/2020)
+-  bump node version
+---
+## [v2.3.11](https://pro-gallery-2-3-11.surge.sh) (16/12/2020)
+ 
+#### GALLERY
+ -  feature (slideshowView): add the option to pass navArrows render function
+
+#### OTHER
+ -  fix (layoutFixer): use react class instead of function
+ - This reverts commit 7f6bd4ccd42d37b0c15291b7f818e36c6e3a6953.
+ -  bump node version in .nvmrc files
+ -  fix (loadMoreAmount): fix functionality in the playground
+ -  fix (layoutFixer): use react class instead of function
+
+---
+## [v2.3.10](https://pro-gallery-2-3-10.surge.sh) (15/12/2020)
+ 
+#### GALLERY
+ -  videos will load after first interaction. fix streched posters
+
+---
 ## [v2.3.9](https://pro-gallery-2-3-9.surge.sh) (14/12/2020)
  
 #### GALLERY